requirements/prod.txt
-r base.txt

gunicorn>=22,<24
uvicorn[standard]>=0.30,<1
django-storages>=1.14,<2
boto3>=1.34,<2
sentry-sdk[django]>=2,<3
django-cloudinary-storage>=0.3,<1
cloudinary>=1.40,<2
